update the pydantic models of the GenomicRegionGenerator pipeline to new format used in other pipelines and change the YAML slightly so that it's easier to use discriminated unions for the source parameters
