+++ data/configs/genomic_region_generator_custom.yaml
@@ -2,18 +2,21 @@
 ### BASIC PARAMETERS ###
 #######################
 
+schema_version: 2
+
 ### General parameters
 dir_output: output_genomic_region_generator_custom # name of the directory where the output files will be written
 
 ### Parameters for genome and gene annotation
-source: custom # required: indicate that own annotation should be used
-source_params:
-  file_annotation: data/annotations/custom_GCF_000001405.40_GRCh38.p14_genomic_chr16.gtf # required: GTF file with gene annotation
-  file_sequence: data/annotations/custom_GCF_000001405.40_GRCh38.p14_genomic_chr16.fna # required: FASTA file with genome sequence
-  files_source: NCBI # optional: original source of the genomic files
-  species: Homo_sapiens # optional: species of provided annotation, leave empty if unknown
-  annotation_release: "110" # optional: release number of provided annotation, leave empty if unknown
-  genome_assembly: GRCh38 # optional: genome assembly of provided annotation, leave empty if unknown
+source:
+  source: custom # required: indicate that own annotation should be used
+  parameters:
+    file_annotation: data/annotations/custom_GCF_000001405.40_GRCh38.p14_genomic_chr16.gtf # required: GTF file with gene annotation
+    file_sequence: data/annotations/custom_GCF_000001405.40_GRCh38.p14_genomic_chr16.fna # required: FASTA file with genome sequence
+    files_source: NCBI # optional: original source of the genomic files
+    species: Homo_sapiens # optional: species of provided annotation, leave empty if unknown
+    annotation_release: "110" # optional: release number of provided annotation, leave empty if unknown
+    genome_assembly: GRCh38 # optional: genome assembly of provided annotation, leave empty if unknown
 
 ### Parameters for sequences generation
 # List of genomic regions that should be generated, set the genomic regions you want to generate to true
