+++ data/configs/genomic_region_generator_ensemble.yaml
@@ -2,14 +2,17 @@
 ### BASIC PARAMETERS ###
 #######################
 
+schema_version: 2
+
 ### General parameters
 dir_output: output_genomic_region_generator_ensembl # name of the directory where the output files will be written
 
 ### Parameters for genome and gene annotation
-source: ensembl # required: indicate that ensembl annotation should be used
-source_params:
-  species: homo_sapiens # required: species name in ensemble download format, e.g. 'homo_sapiens' for human; see http://ftp.ensembl.org/pub/release-108/gtf/ for available species names
-  annotation_release: current # required: release number of annotation, e.g. 'release-108' or 'current' to use most recent annotation release. Check out release numbers for ensemble at ftp.ensembl.org/pub/
+source:
+  source: ensembl # required: indicate that ensembl annotation should be used
+  parameters:
+    species: homo_sapiens # required: species name in ensemble download format, e.g. 'homo_sapiens' for human; see http://ftp.ensembl.org/pub/release-108/gtf/ for available species names
+    annotation_release: current # required: release number of annotation, e.g. 'release-108' or 'current' to use most recent annotation release. Check out release numbers for ensemble at ftp.ensembl.org/pub/
 
 ### Parameters for sequences generation
 # List of genomic regions that should be generated, set the genomic regions you want to generate to true
