+++ data/configs/genomic_region_generator_ncbi.yaml
@@ -2,15 +2,18 @@
 ### BASIC PARAMETERS ###
 #######################
 
+schema_version: 2
+
 ### General parameters
 dir_output: output_genomic_region_generator_ncbi # name of the directory where the output files will be written
 
 ### Parameters for genome and gene annotation
-source: ncbi # required: indicate that ncbi annotation should be used
-source_params:
-  taxon: vertebrate_mammalian # required: taxon of the species, valid taxa are: archaea, bacteria, fungi, invertebrate, mitochondrion, plant, plasmid, plastid, protozoa, vertebrate_mammalian, vertebrate_other, viral
-  species: Homo_sapiens # required: species name in NCBI download format, e.g. 'Homo_sapiens' for human; see https://ftp.ncbi.nlm.nih.gov/genomes/refseq/ for available species name
-  annotation_release: "110" # required: release number of annotation e.g. '109' or '109.20211119'  or 'current' to use most recent annotation release. Check out release numbers for NCBI at ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/annotation/annotation_releases/
+source:
+  source: ncbi # required: indicate that ncbi annotation should be used
+  parameters:
+    taxon: vertebrate_mammalian # required: taxon of the species, valid taxa are: archaea, bacteria, fungi, invertebrate, mitochondrion, plant, plasmid, plastid, protozoa, vertebrate_mammalian, vertebrate_other, viral
+    species: Homo_sapiens # required: species name in NCBI download format, e.g. 'Homo_sapiens' for human; see https://ftp.ncbi.nlm.nih.gov/genomes/refseq/ for available species name
+    annotation_release: "110" # required: release number of annotation e.g. '109' or '109.20211119'  or 'current' to use most recent annotation release. Check out release numbers for NCBI at ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/annotation/annotation_releases/
 
 ### Parameters for sequences generation
 # List of genomic regions that should be generated, set the genomic regions you want to generate to true
